+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog
 1.0.6 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Add a missing profile dependencies
+  [mpeeters]
 
 
 1.0.5 (2019-02-11)