+++ CHANGES.rst
@@ -5,6 +5,12 @@ Changelog
 1.0.6 (unreleased)
 ------------------
 
+- Add tests for creation folder adapter
+  [mpeeters]
+
+- Improve retrievement of creation folder to avoid errors
+  [mpeeters]
+
 - Add a view to ensure that security and permissions are correctly defined for citizen related folders
   [mpeeters]